move EngineChannel classes to their own folder
