+++ src/test/mockedenginebackendtest.h
@@ -12,8 +12,8 @@
 #include "effects/effectsmanager.h"
 #include "engine/enginebuffer.h"
 #include "engine/enginebufferscale.h"
-#include "engine/enginechannel.h"
-#include "engine/enginedeck.h"
+#include "engine/channels/enginechannel.h"
+#include "engine/channels/enginedeck.h"
 #include "engine/enginemaster.h"
 #include "engine/controls/ratecontrol.h"
 #include "engine/sync/enginesync.h"
